file naming, warnings for overwriting files

+++ reshape_mrois_to_planes.py
@@ -89,27 +89,27 @@ mroi_sizes_px = np.array([r['size_px'] for r in md['mrois']['lrsort']], dtype=in
 mroi_corners_tl_px = np.array([r['corner_tl_px'] for r in md['mrois']['lrsort']], dtype=int)
 
 volume = np.full((n_f, n_x, n_y, n_z), np.nan, dtype=np.float32)
+overlap_px = 0
 for i_plane in range(n_z):
-    overlap_seams_this_plane = 0
-    plane_w = n_x - (overlap_seams_this_plane * (md['n_mrois'] - 1))
+    plane_w = n_x - (overlap_px * (md['n_mrois'] - 1))
     plane_h = n_y
     canvas = np.zeros((n_f, plane_w, plane_h), dtype=np.float32)
     for i_mroi in range(md['n_mrois']):
         if i_mroi == 0:
             xs_canv = 0
-            xe_canv = xs_canv + mroi_sizes_px[i_mroi, 0] - int(np.floor(overlap_seams_this_plane / 2))
+            xe_canv = xs_canv + mroi_sizes_px[i_mroi, 0] - int(np.floor(overlap_px / 2))
             xs_mroi = xs_canv
             xe_mroi = xe_canv
         elif i_mroi != md['n_mrois'] - 1:
             xs_canv = xe_canv
-            xe_canv = xs_canv + mroi_sizes_px[i_mroi, 0] - overlap_seams_this_plane
-            x_mroi_width = mroi_sizes_px[i_mroi][0] - overlap_seams_this_plane
-            xs_mroi = int(np.ceil(overlap_seams_this_plane / 2))
+            xe_canv = xs_canv + mroi_sizes_px[i_mroi, 0] - overlap_px
+            x_mroi_width = mroi_sizes_px[i_mroi][0] - overlap_px
+            xs_mroi = int(np.ceil(overlap_px / 2))
             xe_mroi = xs_mroi + x_mroi_width
         else:
             xs_canv = xe_canv
             xe_canv = plane_w
-            xs_mroi = np.ceil(overlap_seams_this_plane / 2).astype(int)
+            xs_mroi = np.ceil(overlap_px / 2).astype(int)
             xe_mroi = mroi_sizes_px[i_mroi][0]
         # print(xs_canv, xe_canv, xs_mroi, xe_mroi)
 
@@ -134,9 +134,12 @@ volume = volume.astype(np.int16)
 volume = np.swapaxes(volume, 1, 2)
 
 md_str = json.dumps(md, default=json_serializer)
+sp = source_path + os.path.sep
 
 if p['save']['hdf5']:
-    save_path_h5 = source_path + os.path.sep + source_name + '_preprocessed.h5'
+    save_path_h5 = sp + source_name + '_preprocd_seamolap{:02d}px.h5'.format(int(overlap_px))
+    if os.path.isfile(save_path_h5):
+        warn('Preprocessed HDF5 ouput already exists, overwriting ({}).'.format(save_path_h5))
     h5f = h5py.File(save_path_h5, 'w')
     h5f.create_dataset('data', data=volume)
     h5f.attrs.create('metadata', str(md_str))
@@ -144,27 +147,42 @@ if p['save']['hdf5']:
     del h5f
 
 if p['save']['tif']:
+    save_path_tif = sp + source_name + '_preprocd_seamolap{:02d}px.tif'.format(int(overlap_px))
+    if os.path.isfile(save_path_tif):
+        warn('Preprocessed TIF ouput already exists, overwriting ({}).'.format(save_path_tif))
+
     import tifffile
-    save_path_tif = source_path + os.path.sep + source_name + '_preprocessed.tif'
+
     tifffile.imwrite(save_path_tif, volume, description=md_str)
 
 if p['save']['metadata']:
-    save_path_md = source_path + os.path.sep + source_name + '_metadata.json'
-    with open(save_path_md, 'w') as mdf:
-        json.dump(md, mdf, indent=4, sort_keys=True, default=json_serializer)
+    save_path_md = sp + source_name + '_metadata.json'
+    if not os.path.isfile(save_path_md):
+        with open(save_path_md, 'w') as mdf:
+            json.dump(md, mdf, indent=4, sort_keys=True, default=json_serializer)
+    else:
+        warn('Metadata file already exists, not overwriting ({}).'.format(save_path_md))
 
 if p['save']['mean']:
+    save_path_mean = sp + source_name + '_preprocd_mean.png'
+    if os.path.isfile(save_path_mean):
+        warn('Preprocessed mean image already exists, overwriting ({}).'.format(save_path_mean))
+
     from cv2 import imwrite
     from skimage.exposure import rescale_intensity
     from skimage.io import imsave
     from skimage.util import img_as_ubyte
-    save_path_mean = source_path + os.path.sep + source_name + '_preprocessed_mean.png'
+
     volume_mean = np.mean(volume, axis=0)
     pl, ph = np.percentile(volume_mean, [1, 99.9])
     volume_mean_rescale = img_as_ubyte(rescale_intensity(volume_mean, in_range=(pl, ph)))
     imwrite(save_path_mean, volume_mean_rescale)
 
 if p['save']['video']:
+    save_path_video = sp + source_name + '_preprocd_seamolap{:02d}px_clip.mp4'.format(int(overlap_px))
+    if os.path.isfile(save_path_video):
+        warn('Video clip already exists, overwriting ({}).'.format(save_path_video))
+
     from cv2 import VideoWriter_fourcc, VideoWriter
     from skimage.exposure import rescale_intensity
     from skimage.io import imsave
@@ -180,9 +198,7 @@ if p['save']['video']:
         ratio_h = h / video_max_h
         ratio_max = max(ratio_w, ratio_h)
         if ratio_max > 1:
-            m = 'Frame size exceeds the maximum allowed by the codec. ' + \
-                'The video output will be downsampled.'
-            warn(m)
+            warn('Frame size exceeds the codec maximum. The video output will be downsampled.')
             dw = round(w / ratio_max)
             dh = round(h / ratio_max)
             print('resizing to w={} h={}'.format(dw, dh))
@@ -201,7 +217,6 @@ if p['save']['video']:
             video_writer.write(stack[f])
         video_writer.release()
 
-    save_path_video = source_path + os.path.sep + source_name + '_preprocessed_clip.mp4'
     n_f = np.min([np.ceil(30 * md['framerate']).astype(int), volume.shape[0]])
     v = volume[0:n_f]
     pl, ph = np.percentile(v, [1, 99.9])
